core: refactor: bridge objects couldn't shares subs
assume that we make a bridge object with bridger.
that object should provide a 'ctor' for managed space.
when an user calls ctor for the bridged object, however, the subs which
belonged to origin bridged object couldn't be shared to all copied
bridged objects.

to solve this, I put all bridged funcs into static variable.
and let bridge class redirects subs to static one.