core: refactor: optimize performance of retConvergence
former convergence put request into queue for all kind of getExpr.
so queue will be going to huge when you writes a thousands of loc.

this new convergence feature is about optimization for reducing pushing a request
into queue.
when expand visits getExpr instance, it make convergene instance and let
it try to converge.
if it's success, symbol has refered by getExpr was already expanded. so,
converge type to real symbol and request will vanish automatically.
if it's not, symbol has not expanded yet. in this case the request will
be going to put into the queue.
expander will runs all requests after successful expansion.
at that time, every requests have only one chance to try.
if convergence is failed again, it means that you write wrong codes.