core: fix: I shouldn't apply orRet when I define tstr<T>
rhs value can be assigned well and orRet will stack up '__orRetStack__'
to 'true'.
however tstr<T> can reject to be assigned the value if its type doesn't
fit to T.
e.g.
    tstr<nStr> value = getRunExpr().run() orRet;
    value->get(); // this code is not safe, actually.

in above example, runExpr::run() returns `str`.
orRet macro catches 'str' value and check whether it's nullable or not.
and definitely str has some value in this step. so, __orRetStack__
pushes 'true' value.
it was 'true' so orRet won't return this function, so codes keep
running.
however, when the output of executing `getRunExpr().run()` to be
assigned to `tstr<nStr>` constructor, it can reject to accept the value,
if the output wasn't derived type of nStr, for instance, if it's nInt.
but code will continue to be executed. orRet macro was evaluated already.

eventually, when program starts to dereferencing of `value` variable,
it goes into NPE.