core: fix(smultimap): app crashes when iterating over `key`
it was an issue that was a bit difficult to find.
it does not occur in DBG and only occurs in REL (i.e. -O2 or higher).
this occurs because the iterator instance of smultimap continues to own the key (std::string type)
created externally as a temporary object in the form of K*(std::string*).
if it is converted to a temporary object and an iterator is created by calling a function as follows,
smultimap<std::string, int> s;
auto e = s.begin("apple"); // char* --> std::string()
....
++e; // crash!

it's fine right away, but if you try to access the key value later, -O2 discards the value for
optimization, so the app dies.