fix(builder): some binary must return error code 1 if it exists
the algorithm for checking dependency was changed, but there was a defect.
this changed algorithm is an improved version that checks in detail whether each binary exists and
whether it satisfies the version information.
the problem is that some binaries do not support the option `--version`, which is usually used to
get version, and also return error code 1, which causes an exception in python's
`subprocess.check_output()`.

So, if an error occurs after running a command to get the version for a binary, it does not
must mean that the binary does not exist. to compensate for this, after an error is returned,
I run `shutil.which()` to additionally check if the binary is indeed installed.