leaf: refactor: not to use nullable reference
fix to avoid using nullable references.

nullable references actually behave as UB, which is not specified in the standard.
there was a clear benefit in terms of code readability that brought in nullable references,
but there was still a risk that UB could cause errors.

it has been maintained for 4-5 years and has been working without any problems in gcc, clang,
msbuild, etc., but this was just luck.
and recently, when compiling with clang on a MacBook with an Apple Silicon chip, a crash occurred
while handling nullable references.
I can't get shot twice after being shot once.

I'm going to remove all the logic that creates all nullable references, checks whether these
references are null and momentarily disables optimization options in the process.
I'm going to handle references within the scope defined by the C++ standard.

unlike other modern OOPs, C++ requires several types to create and manage instances.
in the past, references also played the role of pointers, so only three types were used.

    1) T&
       the most common type for referencing variables. it can be null,
       and can refer to local variables, variables on the heap,
       regardless of the type.

    2) T
       used when returning something as a value from some functions.
       or used when passing it to a function as a temporary argument.

    3) tstr or tweak
       it should be returned as T* or T&, and it is used when you want to manage the lifecycle of
       the instance you are referencing.
       if you simply return it as T*, the instance created on the heap with `new` inside the
       function will not be deleted, which will cause a memory leak.
       these classes use reference counting like unique_ptr to release memory at the appropriate time.
       the reason why unique_ptr is not used is because the memlite module manages memory using
       its own memory pool.

The use case above now changes to:

    1) T&
       used when returning a reference to a member variable held as a value inside a class.
       this reference is compliant with the C++ standard, so it will never be null.

    2) T*
       a function returns a reference to a specific instance, but there are cases where the
       instance does not exist. It may be nullptr, so the function caller must check for null.

    3) T
       used when returning something as a value from some functions.
       or used when passing it to a function as a temporary argument.

    4) tstr or tweak
       it should be returned as T*, and it is used when you want to manage the lifecycle of the
       instance you are referencing.
       if you simply return as T*, the instance created on the heap with `new` inside the function
       will not be deleted, which will cause a memory leak.
       these classes use reference counting like unique_ptr to release memory at the appropriate time.
       the reason why unique_ptr is not used is because the memlite module manages memory using
       its own memory pool.

    5) tmay or tres
       use this when you need to return something as a value, but also need to pass along the success
       or failure of the operation.
       you can check whether it has a normal value with `operator bool()` or `has()`. In this case,
       you cannot get the value with `get()` or `operator*()`.
       if there is no value, you should never call `get()` or `operator*()`. This will result in UB.

       the reason why it has a somewhat heterogeneous concept, such as
       `User check first before calling`, is because tmay or tres are based on stl.

in stl, operator*() is always returned in the form of T& when calling, and if the value is not valid,
the app becomes UB. It usually crashes.

although it is not desirable, namu also needs to be based on STL, so we want to keep this convention.
except for some classes, `get()` returns T*.
so when calling `operator[]()`, you should check for errors in advance, or use it only when you
are sure that it is not an error.