indep: refactor(tmedium): don't take as T& with OR MACRO if a func returns as value
if the function returns a value not a reference or pointer like `str`,
it should not be received as tmedium in the OR macro.

if it is received as tmedium, it will try to resolve it with T&&,
in this case, the original object must be extended by the compiler.
but this is not supported according to the standard document of c++.

as a result, it will receive T&& that refers to an object that has
already been destroyed so the subsequent operation will be UB