core: fix(frameInteractable): remove `args` parameter for `outFrame()`
func::inFrame() needs to receive args because it needs to put arguments into the frame.
if the number of args is greater than 0, args will be added to the current frame that entered
this way as a separate scope.
this means that func::outFrame() needs to know how many scopes to remove.
so outFrame() received args. (actually, you only need to know whether you put arguments in or not,
but for the sake of type consistency, you put them in as args, which is the same type.)