fix(doc): `@lang` annotation conflicts on guide page
When syntax highlighting code examples, they are parsed as JSON by default.
If only code is entered without JSON, it is considered empty JSON.
In this case, the shown and codes values will be the same, and the lang
value will be fixed to byeol.

However, if only the @lang annotation is added, it is assumed to be used on the
reference page, the `verified` class will be added, and the lang value given by
@lang will be used.

If a runnable class is not specified, the `play` button will not appear.